src/main/java/com/tdd/service/CommentService.java
package com.tdd.service;


import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.tdd.domain.Comment;
import com.tdd.domain.CommentRepository;

@Service
public class CommentService {

    private final CommentRepository commentRepository;

    @Autowired
    public CommentService(CommentRepository commentRepository) {
        this.commentRepository = commentRepository;
    }

    public void save(Comment request) {
        commentRepository.save(request);
    }
}
